Added a intermediate table to synceable data
